feat(iam): rename 'manageTokens' to 'permanentTokens'
ref: #MAIAM-45
