+++ packages/manager/apps/identity-access-management/public/translations/permanent-tokens/Messages_fr_FR.json
