+++ packages/manager/apps/identity-access-management/src/routes/routes.constant.ts
@@ -3,7 +3,7 @@ export const subRoutes = {
   assignTag: 'assign-tag',
   tagDetail: ':tag',
   unassignTag: 'unassign-tag',
-  manageTokens: 'manage-tokens/:userId',
+  permanentTokens: 'manage-tokens/:userId',
 };
 
 export const urls = {
@@ -13,5 +13,5 @@ export const urls = {
   tagDetail: `/${subRoutes.tagManager}/${subRoutes.tagDetail}`,
   tagDetailAssign: `/${subRoutes.tagManager}/${subRoutes.tagDetail}/${subRoutes.assignTag}`,
   tagdetailUnassign: `/${subRoutes.tagManager}/${subRoutes.tagDetail}/${subRoutes.unassignTag}`,
-  manageTokens: `/${subRoutes.manageTokens}`,
+  permanentTokens: `/${subRoutes.permanentTokens}`,
 };
